rebase with master and move rpc to separete package
