+++ go.mod
@@ -5,8 +5,8 @@ go 1.15
 require github.com/urfave/cli v1.22.4
 
 require github.com/stretchr/testify v1.6.1
+
 require (
 	github.com/dchest/uniuri v0.0.0-20200228104902-7aecb25e1fe5
-	github.com/stretchr/testify v1.6.1
 	github.com/thanhpk/randstr v1.0.4
 )
