+++ internal/kademlia/kademlia.go
@@ -5,12 +5,12 @@ type Kademlia struct {
 	network Network
 }
 
-//InitNode initializes the Kademlia Node
+// InitNode initializes the Kademlia Node
 // with a Routing Table and a Network
 func (kademlia *Kademlia) InitNode(id *KademliaID) {
 	kademlia.network = Network{}
 	ip := kademlia.network.GetLocalIP()
-	kademlia.network.InitNetwork(ip, "8080")
+	kademlia.network.Listen(ip, "8080")
 
 	me := NewContact(id, ip)
 	kademlia.RT = NewRoutingTable(me)
