+++ internal/rpc/rpc_test.go
@@ -1,4 +1,4 @@
-package network
+package rpc
 
 import (
 	"testing"
